sales-service/requirements-dev.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
pymongo==4.6.1
motor==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0 